[Bronze V] Title: 이상한 기호, Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cc
@@ -8,7 +8,7 @@ typedef vector<int> vi;
 
 
 int main(){
-    int n,m; cin >> n >> m;
+    ll n,m; cin >> n >> m;
     cout << (n+m)*(n-m);
     return 0;
 }
